backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic[email]==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.17
alembic==1.13.3
starlette==0.41.3
itsdangerous==2.2.0
